Refactor before adding support for "embedded db".
