+++ common/actively_persisted_item_test.go
@@ -1,4 +1,4 @@
-package in_red_ck
+package common
 
 import (
 	"testing"
@@ -6,7 +6,7 @@ import (
 
 	"github.com/SharedCode/sop"
 	cas "github.com/SharedCode/sop/cassandra"
-	"github.com/SharedCode/sop/in_red_ck/mocks"
+	"github.com/SharedCode/sop/common/mocks"
 )
 
 func Test_StreamingDataStoreRollbackShouldEraseTIDLogs(t *testing.T) {
@@ -50,7 +50,7 @@ func Test_StreamingDataStoreAbandonedTransactionLogsGetCleaned(t *testing.T) {
 	yesterday := time.Now().Add(time.Duration(-24 * time.Hour))
 	cas.Now = func() time.Time { return yesterday }
 	sop.Now = func() time.Time { return yesterday }
-	Now = func() time.Time { return yesterday }
+	//Now = func() time.Time { return yesterday }
 
 	trans, _ := newMockTransactionWithLogging(t, sop.ForWriting, -1)
 	trans.Begin()
@@ -83,7 +83,7 @@ func Test_StreamingDataStoreAbandonedTransactionLogsGetCleaned(t *testing.T) {
 	today := time.Now()
 	cas.Now = func() time.Time { return today }
 	sop.Now = func() time.Time { return today }
-	Now = func() time.Time { return today }
+	//Now = func() time.Time { return today }
 
 	tid, _, _, _ = twoPhaseTrans.logger.logger.GetOne(ctx)
 	if tid.IsNil() {
