+++ common/btree_with_induced_errors_test.go
@@ -1,4 +1,4 @@
-package in_red_ck
+package common
 
 import (
 	"context"
