+++ common/btree_with_transaction_test.go
@@ -1,4 +1,4 @@
-package in_red_ck
+package common
 
 import (
 	"cmp"
