+++ common/store_interfaces.go
@@ -1,4 +1,4 @@
-package in_red_ck
+package common
 
 import (
 	"github.com/SharedCode/sop/btree"
