+++ in_red_ck/integration_tests/delete_btree_test.go
@@ -8,7 +8,8 @@ import (
 
 // Add Test_ prefix if you want to run this test.
 // It drops the blob & registry tables of the B-Tree, thus, the test was removed from the set.
-func DeleteBTree(t *testing.T) {
+func TestDeleteBTree(t *testing.T) {
+	in_red_ck.Initialize(cassConfig, redisConfig)
 	tableList := []string{
 		"fooStore", "persondb", "twophase", "twophase2", "twophase3",
 		"twophase22", "persondb7", "persondb77", "person2db", "barStore1",
