+++ in_red_ck/integration_tests/transaction_logging_test.go
@@ -17,7 +17,6 @@ func MultipleExpiredTransCleanup(t *testing.T) {
 	yesterday := time.Now().Add(time.Duration(-48 * time.Hour))
 	cas.Now = func() time.Time { return yesterday }
 	sop.Now = func() time.Time { return yesterday }
-	in_red_ck.Now = func() time.Time { return yesterday }
 
 	trans, _ := in_red_ck.NewTransaction(sop.ForWriting, -1, true)
 	trans.Begin()
@@ -42,7 +41,6 @@ func MultipleExpiredTransCleanup(t *testing.T) {
 	yesterday = time.Now().Add(time.Duration(-47 * time.Hour))
 	cas.Now = func() time.Time { return yesterday }
 	sop.Now = func() time.Time { return yesterday }
-	in_red_ck.Now = func() time.Time { return yesterday }
 
 	trans, _ = in_red_ck.NewTransaction(sop.ForWriting, -1, true)
 	trans.Begin()
@@ -57,7 +55,6 @@ func MultipleExpiredTransCleanup(t *testing.T) {
 	yesterday = time.Now().Add(time.Duration(-46 * time.Hour))
 	cas.Now = func() time.Time { return yesterday }
 	sop.Now = func() time.Time { return yesterday }
-	in_red_ck.Now = func() time.Time { return yesterday }
 
 	trans, _ = in_red_ck.NewTransaction(sop.ForWriting, -1, true)
 	trans.Begin()
@@ -71,7 +68,6 @@ func MultipleExpiredTransCleanup(t *testing.T) {
 	yesterday = time.Now()
 	cas.Now = func() time.Time { return yesterday }
 	sop.Now = func() time.Time { return yesterday }
-	in_red_ck.Now = func() time.Time { return yesterday }
 
 	trans, _ = in_red_ck.NewTransaction(sop.ForWriting, -1, true)
 
@@ -84,7 +80,6 @@ func Cleanup(t *testing.T) {
 	yesterday := time.Now().Add(time.Duration(-24 * time.Hour))
 	cas.Now = func() time.Time { return yesterday }
 	sop.Now = func() time.Time { return yesterday }
-	in_red_ck.Now = func() time.Time { return yesterday }
 
 	trans, _ := in_red_ck.NewTransaction(sop.ForReading, -1, true)
 	trans.Begin()
@@ -94,7 +89,6 @@ func Cleanup(t *testing.T) {
 	yesterday = time.Now().Add(-time.Duration(23*time.Hour + 54*time.Minute))
 	cas.Now = func() time.Time { return yesterday }
 	sop.Now = func() time.Time { return yesterday }
-	in_red_ck.Now = func() time.Time { return yesterday }
 
 	trans, _ = in_red_ck.NewTransaction(sop.ForReading, -1, true)
 	trans.Begin()
