+++ in_red_cs3/integration_tests/transaction_logging_test.go
@@ -7,7 +7,6 @@ import (
 
 	"github.com/SharedCode/sop"
 	cas "github.com/SharedCode/sop/cassandra"
-	"github.com/SharedCode/sop/in_red_ck"
 	"github.com/SharedCode/sop/in_red_cs3"
 )
 
@@ -22,7 +21,6 @@ func MultipleExpiredTransCleanup(t *testing.T) {
 	yesterday := time.Now().Add(time.Duration(-48 * time.Hour))
 	cas.Now = func() time.Time { return yesterday }
 	sop.Now = func() time.Time { return yesterday }
-	in_red_ck.Now = func() time.Time { return yesterday }
 
 	trans, _ := in_red_cs3.NewTransaction(s3Client, sop.ForWriting, -1, true, region)
 	trans.Begin()
@@ -47,7 +45,6 @@ func MultipleExpiredTransCleanup(t *testing.T) {
 	yesterday = time.Now().Add(time.Duration(-47 * time.Hour))
 	cas.Now = func() time.Time { return yesterday }
 	sop.Now = func() time.Time { return yesterday }
-	in_red_ck.Now = func() time.Time { return yesterday }
 
 	trans, _ = in_red_cs3.NewTransaction(s3Client, sop.ForWriting, -1, true, region)
 	trans.Begin()
@@ -62,7 +59,6 @@ func MultipleExpiredTransCleanup(t *testing.T) {
 	yesterday = time.Now().Add(time.Duration(-46 * time.Hour))
 	cas.Now = func() time.Time { return yesterday }
 	sop.Now = func() time.Time { return yesterday }
-	in_red_ck.Now = func() time.Time { return yesterday }
 
 	trans, _ = in_red_cs3.NewTransaction(s3Client, sop.ForWriting, -1, true, region)
 	trans.Begin()
@@ -76,7 +72,6 @@ func MultipleExpiredTransCleanup(t *testing.T) {
 	yesterday = time.Now()
 	cas.Now = func() time.Time { return yesterday }
 	sop.Now = func() time.Time { return yesterday }
-	in_red_ck.Now = func() time.Time { return yesterday }
 
 	trans, _ = in_red_cs3.NewTransaction(s3Client, sop.ForWriting, -1, true, region)
 
@@ -89,7 +84,6 @@ func Cleanup(t *testing.T) {
 	yesterday := time.Now().Add(time.Duration(-24 * time.Hour))
 	cas.Now = func() time.Time { return yesterday }
 	sop.Now = func() time.Time { return yesterday }
-	in_red_ck.Now = func() time.Time { return yesterday }
 
 	trans, _ := in_red_cs3.NewTransaction(s3Client, sop.ForReading, -1, true, region)
 	trans.Begin()
@@ -99,7 +93,6 @@ func Cleanup(t *testing.T) {
 	yesterday = time.Now().Add(-time.Duration(23*time.Hour + 54*time.Minute))
 	cas.Now = func() time.Time { return yesterday }
 	sop.Now = func() time.Time { return yesterday }
-	in_red_ck.Now = func() time.Time { return yesterday }
 
 	trans, _ = in_red_cs3.NewTransaction(s3Client, sop.ForReading, -1, true, region)
 	trans.Begin()
