+++ redis/redis_test.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 	"testing"
-	// "golang.org/x/sync/errgroup"
 )
 
 type user struct {
